# Reduced allocated memory by chunks reader in the store gateway at query time