--- a/pkg/store/io.go
+++ b/pkg/store/io.go
@@ -66,9 +66,9 @@ func readByteRanges(src io.Reader, dst []byte, byteRanges byteRanges) ([]byte, e
 
 	// Optimisation for the case all ranges are contiguous.
 	if byteRanges[0].offset == 0 && byteRanges.areContiguous() {
-		// We get an ErrUnexpectedEOF if EOF is reached after some but not all bytes
-		// have been read. Due to how the reading logic works in the bucket store, we
-		// may try to overread at the end of an object, so we consider it legit.
+		// We get an ErrUnexpectedEOF if EOF is reached before we fill allocated dst slice.
+		// Due to how the reading logic works in the bucket store, we may try to overread at
+		// the end of an object, so we consider it legit.
 		if _, err := io.ReadFull(src, dst); err != nil && err != io.ErrUnexpectedEOF {
 			return nil, err
 		}
@@ -107,10 +107,9 @@ func readByteRanges(src io.Reader, dst []byte, byteRanges byteRanges) ([]byte, e
 			dstOffset += readBytes
 		}
 		if err != nil {
-			// We get an ErrUnexpectedEOF if EOF is reached after some but not all bytes
-			// have been read. Due to how the reading logic works in the bucket store, we
-			// may try to overread the last byte range so, if the error occurrs on the last
-			// one, we consider it legit.
+			// We get an ErrUnexpectedEOF if EOF is reached before we fill the slice.
+			// Due to how the reading logic works in the bucket store, we may try to overread
+			// the last byte range so, if the error occurrs on the last one, we consider it legit.
 			if err == io.ErrUnexpectedEOF && idx == len(byteRanges)-1 {
 				return dst, nil
 			}