--- a/pkg/store/io_test.go
+++ b/pkg/store/io_test.go
@@ -1,10 +1,14 @@
+// Copyright (c) The Thanos Authors.
+// Licensed under the Apache License 2.0.
+
 package store
 
 import (
 	"bytes"
-	"errors"
 	"io"
 	"testing"
+
+	"github.com/pkg/errors"
 
 	"github.com/thanos-io/thanos/pkg/testutil"
 )