* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

## Changes

Cortex store-gateway (based on Thanos one) allocates a lot of memory at query time. After some debugging on some high cardinality queries, we've noticed most of the memory being allocated to read chunks.

At a first glance, there may appear there's nothing we can do to reduce the memory to read chunks (all in all we have to read them) but we noticed a good % of memory being wasted due to the partitioner.

The partitioner is a smart and good idea to reduce the number of bucket API calls and we don't want to sacrifice it, but we don't even need to retain in memory all bytes overread by the partitioner which don't belong to actual chunks we have to read. This PR leverages on this and basically trash away all bytes overread by the partitioner without storing them in memory.

I've added a new benchmark to show the improvements and I'm running it with different concurrency because of potential side effects of the memory pool (eg. sequential queries may false the benchmark).

Finally, in this PR I tried to keep the logic as simplest as possible. There are further optimisations we can do (and I plan to do) but I would prefer to work iteratively to simplify the code review process.

## Benchmarks

- `BenchmarkBucketSeries` always query all block series, so no optimisation here (but neither perf degrade)
- `BenchmarkBlockSeries` is the new benchmark I've added (look at allocated memory)

```
benchmark                                                                  old ns/op      new ns/op      delta
BenchmarkBucketSeries/1000000SeriesWith1Samples/1of1000000-12              106619669      107524709      +0.85%
BenchmarkBucketSeries/1000000SeriesWith1Samples/10of1000000-12             106322544      104727465      -1.50%
BenchmarkBucketSeries/1000000SeriesWith1Samples/1000000of1000000-12        1018314234     1009187596     -0.90%
BenchmarkBucketSeries/100000SeriesWith100Samples/1of10000000-12            6131079        6116107        -0.24%
BenchmarkBucketSeries/100000SeriesWith100Samples/100of10000000-12          6148111        6106837        -0.67%
BenchmarkBucketSeries/100000SeriesWith100Samples/10000000of10000000-12     83858366       85699005       +2.19%
BenchmarkBucketSeries/1SeriesWith10000000Samples/1of10000000-12            202194         200119         -1.03%
BenchmarkBucketSeries/1SeriesWith10000000Samples/100of10000000-12          203532         202342         -0.58%
BenchmarkBucketSeries/1SeriesWith10000000Samples/10000000of10000000-12     24225932       24187450       -0.16%
BenchmarkBucketBlock_readChunkRange-12                                     51056          49381          -3.28%
BenchmarkBlockSeries/concurrency:_1-12                                     12662116       12708911       +0.37%
BenchmarkBlockSeries/concurrency:_2-12                                     7590030        7119299        -6.20%
BenchmarkBlockSeries/concurrency:_4-12                                     5657066        4511935        -20.24%
BenchmarkBlockSeries/concurrency:_8-12                                     4943242        3778894        -23.55%
BenchmarkBlockSeries/concurrency:_16-12                                    7420097        4654564        -37.27%
BenchmarkBlockSeries/concurrency:_32-12                                    8118744        6531318        -19.55%

benchmark                                                                  old allocs     new allocs     delta
BenchmarkBucketSeries/1000000SeriesWith1Samples/1of1000000-12              9671           9717           +0.48%
BenchmarkBucketSeries/1000000SeriesWith1Samples/10of1000000-12             9775           9802           +0.28%
BenchmarkBucketSeries/1000000SeriesWith1Samples/1000000of1000000-12        10078124       10077846       -0.00%
BenchmarkBucketSeries/100000SeriesWith100Samples/1of10000000-12            1107           1105           -0.18%
BenchmarkBucketSeries/100000SeriesWith100Samples/100of10000000-12          1143           1134           -0.79%
BenchmarkBucketSeries/100000SeriesWith100Samples/10000000of10000000-12     1008260        1008272        +0.00%
BenchmarkBucketSeries/1SeriesWith10000000Samples/1of10000000-12            208            207            -0.48%
BenchmarkBucketSeries/1SeriesWith10000000Samples/100of10000000-12          208            207            -0.48%
BenchmarkBucketSeries/1SeriesWith10000000Samples/10000000of10000000-12     169743         169733         -0.01%
BenchmarkBucketBlock_readChunkRange-12                                     12             10             -16.67%
BenchmarkBlockSeries/concurrency:_1-12                                     32541          32540          -0.00%
BenchmarkBlockSeries/concurrency:_2-12                                     33244          32054          -3.58%
BenchmarkBlockSeries/concurrency:_4-12                                     35504          33320          -6.15%
BenchmarkBlockSeries/concurrency:_8-12                                     35636          33674          -5.51%
BenchmarkBlockSeries/concurrency:_16-12                                    53859          42102          -21.83%
BenchmarkBlockSeries/concurrency:_32-12                                    64555          61618          -4.55%

benchmark                                                                  old bytes      new bytes      delta
BenchmarkBucketSeries/1000000SeriesWith1Samples/1of1000000-12              62009958       62013891       +0.01%
BenchmarkBucketSeries/1000000SeriesWith1Samples/10of1000000-12             62013629       62019172       +0.01%
BenchmarkBucketSeries/1000000SeriesWith1Samples/1000000of1000000-12        1284131568     1300095624     +1.24%
BenchmarkBucketSeries/100000SeriesWith100Samples/1of10000000-12            4820040        4818485        -0.03%
BenchmarkBucketSeries/100000SeriesWith100Samples/100of10000000-12          4821284        4820303        -0.02%
BenchmarkBucketSeries/100000SeriesWith100Samples/10000000of10000000-12     122767861      131594871      +7.19%
BenchmarkBucketSeries/1SeriesWith10000000Samples/1of10000000-12            179134         177827         -0.73%
BenchmarkBucketSeries/1SeriesWith10000000Samples/100of10000000-12          179264         178000         -0.71%
BenchmarkBucketSeries/1SeriesWith10000000Samples/10000000of10000000-12     42016471       42689769       +1.60%
BenchmarkBucketBlock_readChunkRange-12                                     204766         204521         -0.12%
BenchmarkBlockSeries/concurrency:_1-12                                     28213098       18796949       -33.38%
BenchmarkBlockSeries/concurrency:_2-12                                     29902672       17924383       -40.06%
BenchmarkBlockSeries/concurrency:_4-12                                     30633940       15930915       -48.00%
BenchmarkBlockSeries/concurrency:_8-12                                     29590364       15374578       -48.04%
BenchmarkBlockSeries/concurrency:_16-12                                    44226166       15620776       -64.68%
BenchmarkBlockSeries/concurrency:_32-12                                    48646170       24366022       -49.91%
```

## Verification

Unit tests + benchmarks.