@pracucci 2021-02-22
I'm updating the benchmarks, running them multiple times and comparing with `benchstat`. I've run it with:
```
go test -run '^$' -bench 'BenchmarkBucketSeries|BenchmarkBucketBlock_readChunkRange|BenchmarkBlockSeries' -benchmem -count=5 ./pkg/store/
```

Results:
```
name                                                           old time/op    new time/op    delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-12              106ms ± 2%     105ms ± 2%     ~     (p=0.310 n=5+5)
BucketSeries/1000000SeriesWith1Samples/10of1000000-12             106ms ± 1%     106ms ± 1%     ~     (p=0.690 n=5+5)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-12        971ms ± 3%     956ms ± 3%     ~     (p=0.095 n=5+5)
BucketSeries/100000SeriesWith100Samples/1of10000000-12           6.23ms ± 3%    6.18ms ± 2%     ~     (p=0.151 n=5+5)
BucketSeries/100000SeriesWith100Samples/100of10000000-12         6.20ms ± 1%    6.15ms ± 0%   -0.83%  (p=0.032 n=5+5)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-12    85.8ms ± 1%    85.7ms ± 1%     ~     (p=1.000 n=5+4)
BucketSeries/1SeriesWith10000000Samples/1of10000000-12            206µs ± 2%     206µs ± 1%     ~     (p=0.841 n=5+5)
BucketSeries/1SeriesWith10000000Samples/100of10000000-12          206µs ± 1%     206µs ± 1%     ~     (p=0.690 n=5+5)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-12    24.1ms ± 1%    24.3ms ± 1%     ~     (p=0.095 n=5+5)
BucketBlock_readChunkRange-12                                    51.3µs ± 1%    51.8µs ± 9%     ~     (p=0.690 n=5+5)
BlockSeries/concurrency:_1-12                                    12.4ms ± 2%    12.3ms ± 3%     ~     (p=0.421 n=5+5)
BlockSeries/concurrency:_2-12                                    7.38ms ± 1%    6.92ms ± 2%   -6.22%  (p=0.008 n=5+5)
BlockSeries/concurrency:_4-12                                    5.52ms ± 6%    4.54ms ± 1%  -17.74%  (p=0.008 n=5+5)
BlockSeries/concurrency:_8-12                                    5.20ms ± 2%    3.75ms ± 4%  -27.97%  (p=0.008 n=5+5)
BlockSeries/concurrency:_16-12                                   7.12ms ± 4%    3.51ms ±26%  -50.76%  (p=0.008 n=5+5)
BlockSeries/concurrency:_32-12                                   7.82ms ± 6%    6.13ms ± 4%  -21.64%  (p=0.008 n=5+5)

name                                                           old alloc/op   new alloc/op   delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-12             62.0MB ± 0%    62.0MB ± 0%     ~     (p=0.095 n=5+5)
BucketSeries/1000000SeriesWith1Samples/10of1000000-12            62.0MB ± 0%    62.0MB ± 0%     ~     (p=0.151 n=5+5)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-12       1.27GB ± 1%    1.29GB ± 0%     ~     (p=0.056 n=5+5)
BucketSeries/100000SeriesWith100Samples/1of10000000-12           4.82MB ± 0%    4.82MB ± 0%   -0.05%  (p=0.008 n=5+5)
BucketSeries/100000SeriesWith100Samples/100of10000000-12         4.82MB ± 0%    4.82MB ± 0%     ~     (p=0.056 n=5+5)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-12     127MB ± 7%     129MB ± 3%     ~     (p=0.548 n=5+5)
BucketSeries/1SeriesWith10000000Samples/1of10000000-12            180kB ± 0%     178kB ± 0%   -0.69%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/100of10000000-12          180kB ± 0%     179kB ± 0%   -0.72%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-12    43.4MB ± 5%    44.9MB ± 2%     ~     (p=0.151 n=5+5)
BucketBlock_readChunkRange-12                                     205kB ± 0%     205kB ± 0%   -0.14%  (p=0.008 n=5+5)
BlockSeries/concurrency:_1-12                                    29.4MB ± 5%    17.4MB ± 5%  -40.89%  (p=0.008 n=5+5)
BlockSeries/concurrency:_2-12                                    28.7MB ± 3%    16.9MB ± 6%  -41.29%  (p=0.008 n=5+5)
BlockSeries/concurrency:_4-12                                    29.7MB ± 5%    16.9MB ± 7%  -43.12%  (p=0.008 n=5+5)
BlockSeries/concurrency:_8-12                                    31.3MB ± 3%    15.0MB ± 8%  -52.02%  (p=0.008 n=5+5)
BlockSeries/concurrency:_16-12                                   44.1MB ± 5%    12.4MB ±23%  -71.86%  (p=0.008 n=5+5)
BlockSeries/concurrency:_32-12                                   45.2MB ± 4%    25.4MB ±12%  -43.74%  (p=0.008 n=5+5)

name                                                           old allocs/op  new allocs/op  delta
BucketSeries/1000000SeriesWith1Samples/1of1000000-12              9.70k ± 0%     9.70k ± 0%     ~     (p=0.833 n=5+5)
BucketSeries/1000000SeriesWith1Samples/10of1000000-12             9.79k ± 0%     9.80k ± 0%     ~     (p=0.190 n=5+5)
BucketSeries/1000000SeriesWith1Samples/1000000of1000000-12        10.1M ± 0%     10.1M ± 0%     ~     (p=0.151 n=5+5)
BucketSeries/100000SeriesWith100Samples/1of10000000-12            1.11k ± 0%     1.10k ± 0%     ~     (p=0.079 n=5+5)
BucketSeries/100000SeriesWith100Samples/100of10000000-12          1.14k ± 0%     1.14k ± 0%     ~     (p=0.183 n=5+5)
BucketSeries/100000SeriesWith100Samples/10000000of10000000-12     1.01M ± 0%     1.01M ± 0%     ~     (p=0.500 n=5+5)
BucketSeries/1SeriesWith10000000Samples/1of10000000-12              208 ± 0%       207 ± 0%   -0.48%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/100of10000000-12            208 ± 0%       207 ± 0%   -0.48%  (p=0.008 n=5+5)
BucketSeries/1SeriesWith10000000Samples/10000000of10000000-12      170k ± 0%      170k ± 0%     ~     (p=1.000 n=5+5)
BucketBlock_readChunkRange-12                                      12.0 ± 0%      10.0 ± 0%  -16.67%  (p=0.008 n=5+5)
BlockSeries/concurrency:_1-12                                     32.7k ± 2%     32.0k ± 3%   -2.26%  (p=0.032 n=5+5)
BlockSeries/concurrency:_2-12                                     32.8k ± 1%     31.8k ± 2%   -3.21%  (p=0.008 n=5+5)
BlockSeries/concurrency:_4-12                                     34.9k ± 7%     32.7k ± 0%     ~     (p=0.111 n=5+4)
BlockSeries/concurrency:_8-12                                     37.5k ± 3%     33.5k ± 2%  -10.73%  (p=0.008 n=5+5)
BlockSeries/concurrency:_16-12                                    54.2k ± 2%     33.5k ±21%  -38.29%  (p=0.008 n=5+5)
BlockSeries/concurrency:_32-12                                    62.7k ± 1%     57.8k ± 8%   -7.75%  (p=0.032 n=5+5)
```